% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laterality.R
\docType{data}
\name{aal_paired_regions}
\alias{aal_paired_regions}
\title{Calculate Laterality Index from AAL 116 ROIs}
\format{
An object of class \code{character} of length 54.
}
\usage{
aal_paired_regions
}
\arguments{
\item{x}{\code{data.frame} or similar with columns that include the ALL label 
names ending in "_L" and "_R" for the paired regions}

\item{idname}{Character: Name of column holding cases IDs. Default = "ID"}

\item{select_pattern:}{Character: Select columns that match this pattern}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\description{
Calculate Laterality Index from AAL 116 ROIs
}
\author{
E.D. Gennatas
}
\keyword{datasets}
