% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgmetrics.R
\name{cnr}
\alias{cnr}
\title{Contrast-to-Noise Ratio}
\usage{
cnr(
  x,
  modality = "t1",
  lowThresh = NULL,
  highThresh = NULL,
  cleanup = 2,
  verbose = TRUE
)
}
\arguments{
\item{x}{Character or antsImage: path to image OR ANTs image}

\item{modality}{Character: "t1", "t1v0", "t1nobrainer", "t1combined", "t2", 
"flair", "bold", "fa", "t1t2infant", "t1infant", "t2infant". Default = "t1"
Passed to \code{ANTsRNet::brainExtraction}}

\item{lowThresh}{Double: Passed to \code{ANTsRCore::getMask}}

\item{highThresh}{Double: Passed to \code{ANTsRCore::getMask}}

\item{cleanup}{Double: Passed to \code{ANTsRCore::getMask}. Default = 2}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\value{
List with cnr, snr, img/brain/background mean and sd
}
\description{
CNR = (mean(brain) - mean(background))/sd(background)
}
\author{
E.D. Gennatas
}
