% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelstats.R
\name{labelstats}
\alias{labelstats}
\title{Get label stats for a list of images using \pkg{ANTsR}}
\usage{
labelstats(
  x,
  labeled_nifti,
  labelkey = NULL,
  exclude_label_index = 1,
  verbose = TRUE
)
}
\arguments{
\item{x}{Character vector of nifti paths OR list of \code{antsImage} objects}

\item{labeled_nifti}{Character of path to labeled nifti OR \code{antsImage} object}

\item{labelkey}{Character vector: Label names}

\item{exclude_label_index}{Integer vector: 1-based index of labels to exclude. Default = 1,
which would exclude the first label that is commonly the background. Set to NULL to not exclude
any labels}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\description{
Get label stats for a list of images using \pkg{ANTsR}
}
\author{
E.D. Gennatas
}
