% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgmetrics.R
\name{snr}
\alias{snr}
\title{Signal-to-Noise Ratio}
\usage{
snr(x, verbose = TRUE)
}
\arguments{
\item{x}{Character or antsImage: path to image OR ANTs image}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\value{
List with snr and image mean, sd
}
\description{
SNR = mean(image)/sd(image)
}
\author{
E.D. Gennatas
}
