% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1_dkt.R
\name{t1_dkt}
\alias{t1_dkt}
\title{DKT parcellation using \pkg{ANTsRNet}}
\usage{
t1_dkt(
  x,
  doPreprocessing = TRUE,
  returnProbabilityImages = FALSE,
  writeseg = TRUE,
  preprocdir = file.path(dirname(x), "preprocess_t1"),
  verbose = TRUE
)
}
\arguments{
\item{x}{Character path to T1 file}

\item{doPreprocessing}{Logical: see \code{ANTsRNet::desikanKillianyTourvilleLabeling}}

\item{returnProbabilityImages}{Logical: see \code{ANTsRNet::desikanKillianyTourvilleLabeling}}

\item{writeseg}{Logical: If TRUE, write segmentation to file. Default = TRUE}

\item{preprocdir}{Character: Path to directory to save segmentation, if \code{writeseg = TRUE}}

\item{verbose}{Logical: If TRUE, write messages to console}
}
\description{
Simple wrapper to perform parcellation and write to file.
}
