% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_preprocess.R
\name{which_preprocess}
\alias{which_preprocess}
\title{Check which images have been preprocessed}
\usage{
which_preprocess(
  root_dir,
  raw_suffix = ".nii.gz",
  preprocess_dir = "preprocess_t1",
  target_file_suffix = "rsGM.nii.gz",
  verbose = TRUE
)
}
\arguments{
\item{root_dir}{Character: Directory where raw T1 volumes are stored}

\item{raw_suffix}{Character: Suffix of raw T1 volumes. Default = ".nii.gz"}

\item{preprocess_dir}{Character: Name of preprocessing directory. Default = "preprocess_t1"}

\item{target_file_suffix}{Character: Suffix of preprocessed volumes. Default = "rsGM.nii.gz"}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\description{
Check which images have been preprocessed
}
\author{
E.D. Gennatas
}
