% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelstats_native.R
\name{labelstats_native}
\alias{labelstats_native}
\title{Get label stats for a list of images using \pkg{ANTsR}}
\usage{
labelstats_native(
  x,
  labeled_nifti,
  labelkey = NULL,
  exclude_label_index = NULL,
  do_save_native_label = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{Character vector of nifti paths OR list of \code{antsImage} objects in native space}

\item{labeled_nifti}{Character of path to labeled nifti OR \code{antsImage} object}

\item{labelkey}{Data frame with two columns: "LabelID" (integer) and "LabelName" (character).
If not NULL, output will contain all labels in \code{labelkey} whether they have values or not,
and will exclude any labeld found in \code{labeled_nifti} but not in \code{labelkey}, this
may often be the background label (commonly LabelID 0), for example. Best to include a \code{labelkey}
if available.}

\item{exclude_label_index}{Integer vector: 1-based index of labels to exclude. Default = NULL,
i.e. do not exclude any labels. May be best to include a \code{labelkey} that excludes unwanted
labels (e.g. the background)
any labels}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\description{
Get label stats for a list of images using \pkg{ANTsR}
}
\author{
E.D. Gennatas
}
