% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_t1.R
\name{preprocess_t1}
\alias{preprocess_t1}
\title{Preprocess T1 volume using \pkg{ANTsR} and \pkg{ANTsRNet}}
\usage{
preprocess_t1(
  x,
  id = NULL,
  intensityTruncation = c(0.025, 0.975, 256),
  template_brain = NULL,
  transform_type = "antsRegistrationSyN[s]",
  atropos_initialization = "KMeans[3]",
  atropos_mrf = "[0,1x1x1]",
  atropos_convergence = "[3,0]",
  do_warpgm = FALSE,
  outdir = NULL,
  verbose = TRUE,
  vistrace = 0,
  vistrace_slices = c(90, 130, 170),
  itk_cores = NULL
)
}
\arguments{
\item{x}{Character vector of paths to nifti files OR list of \code{antsImage} volumes}

\item{id}{Character, vector: Names of volumes in \code{x}. Optional; if NULL, defaults to
\code{basename(x)} if \code{x} is character, or names of list}

\item{intensityTruncation}{iMath intensity truncation parameters}

\item{template_brain}{Path to image or antsImage object for template brain, i.e. skull-stripped
volume}

\item{transform_type}{Character: Registration transform type to use. Default = "SyN"}

\item{atropos_initialization}{Character: Atropos initialization. Default = "KMeans[3]"}

\item{atropos_mrf}{Character: Atropos MRF parameters. Default = "[0,1x1x1]" (no smoothing)}

\item{atropos_convergence}{Character: Atropos convergence parameters. Default = "[3,0]"}

\item{outdir}{Character: Path to output directory to save transformations and output volumes}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{vistrace}{Integer 0-2: If > 0, print plots of images at different preprocessing stages}

\item{vistrace_slices}{Integer: How many slices to plot if \code{vistrace > 0}}
}
\description{
Preprocess T1 volume using \pkg{ANTsR} and \pkg{ANTsRNet}
}
\author{
E.D. Gennatas
}
